Adds View Snippet Controller and Partial
In this commit, we update the Recent Snippets partial
(recent-snippets.html) and the Search Results partial
(search-results.html) to include links. The title now appears
as a hyperlink that will navigate to the View Snippet partial
(view-snippet.html).

In addition, we create the ViewSnippet controller and register it
in the app.js file.

Finally, we update content in the Home page and About page.