Updates Snippets Factory with getSnippetById
Adds function getSnippetById to the services.js file.

Adds a call to this function in the ViewSnippetController.

Now, when you click on a snippet title link in the Recent Snippets
view or from the Search Results, you can see the snippet title in
the View Snippet partial. The rest of the view snippet page UI is
still pending.