Adds Language to Search Results and Add Snippet
This commit adds a Language column to the Search Results grid and
begins to implement the Language drop-down-list in the Add a
Snippet Controller.

An ng-change directive has been added to the Language drop-down-list
and it is set to run a function. Inside that function, we still
need to trigger Ace Editor to change its syntax. Working on it...

This commit also adds Lodash to the index.html and as a dependency
in package.json.